apps/backend/pyproject.toml
[project]
name = "saksham-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "google-cloud-aiplatform>=1.87.0",
    "google-cloud-firestore>=2.27.0",
    "google-cloud-translate>=3.26.0",
    "google-cloud-texttospeech>=2.36.0",
    "google-cloud-bigquery>=3.41.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "httpx>=0.28.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "mypy>=1.20.2",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=agents --cov=api --cov=core --cov=services --cov=main --cov-report=term --cov-fail-under=80"

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["google.cloud.*", "vertexai.*", "structlog.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP"]
ignore = ["E501"]
